Adds subscriptions

+++ internal/models/currency.go
@@ -2,8 +2,9 @@ package models
 
 import "time"
 
+// CurrencyRate модель валютной пары
 type CurrencyRate struct {
-	ID        uint      `gorm:"primaryKey"`
+	ID        int64     `gorm:"primaryKey"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime"`
 	Value     float64
 	Name      string
